move files and update SideBar structure
